+++ quantify_scheduler/backends/types/qblox.py
@@ -359,6 +359,8 @@ class SequencerSettings(DataClassJsonMixin):
 
     nco_en: bool
     """Specifies whether the NCO will be used or not."""
+    mix_lo: bool
+    """Specifies whether IQ mixing will be used or not."""
     sync_en: bool
     """Enables party-line synchronization."""
     connected_outputs: Union[Tuple[int], Tuple[int, int]]
@@ -438,6 +440,7 @@ class SequencerSettings(DataClassJsonMixin):
         settings = cls(
             nco_en=nco_en,
             sync_en=True,
+            mix_lo=True,
             connected_outputs=connected_outputs,
             modulation_freq=modulation_freq,
             mixer_corr_gain_ratio=mixer_amp_ratio,